requirements.txt
pandas==1.5.3
fastapi==0.99.0
pillow==9.5.0